Создал таск контроллер
