+++ src/components/day-events-list.js
@@ -1,0 +1,10 @@
+import {AbstractComponent} from "./abstract-component";
+
+export class DayEventsList extends AbstractComponent {
+  constructor() {
+    super();
+  }
+  getTemplate() {
+    return `<ul class="trip-events__list"></ul>`;
+  }
+}
