+++ src/components/trip-day-list.js
@@ -1,0 +1,11 @@
+import {AbstractComponent} from "./abstract-component.js";
+
+export class TripDaysList extends AbstractComponent {
+  constructor() {
+    super();
+  }
+
+  getTemplate() {
+    return `<ul class="trip-days"></ul>`;
+  }
+}
